@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "license": "Apache-2.0",
   "engines": {
-    "node": ">=8.9.0"
+    "node": "10.14.1"
   },
   "scripts": {
     "build": "ng build storefrontapp --prod",
