@@ -2,7 +2,6 @@ import { Action } from '@ngrx/store';
 import { Product } from '../../../occ-models';
 
 export const LOAD_PRODUCT = '[Product] Load Product Data';
-export const LOAD_PRODUCT_START = '[Product] Load Product Data Start';
 export const LOAD_PRODUCT_FAIL = '[Product] Load Product Data Fail';
 export const LOAD_PRODUCT_SUCCESS = '[Product] Load Product Data Success';
 
@@ -16,11 +15,6 @@ export class LoadProduct implements Action {
   }
 }
 
-export class LoadProductStart implements Action {
-  readonly type = LOAD_PRODUCT_START;
-  constructor(public payload: string) {}
-}
-
 export class LoadProductFail implements Action {
   readonly type = LOAD_PRODUCT_FAIL;
   constructor(public payload: any) {}
@@ -32,8 +26,4 @@ export class LoadProductSuccess implements Action {
 }
 
 // action types
-export type ProductAction =
-  | LoadProduct
-  | LoadProductStart
-  | LoadProductFail
-  | LoadProductSuccess;
+export type ProductAction = LoadProduct | LoadProductFail | LoadProductSuccess;
