@@ -10,23 +10,6 @@ export function reducer(
   action: fromProduct.ProductAction
 ): ProductState {
   switch (action.type) {
-    case fromProduct.LOAD_PRODUCT_START: {
-      const code = action.payload;
-      const newState: any = {
-        loading: true
-      };
-      if (state.entities[code]) {
-        newState.value = state.entities[code].value;
-      }
-      return {
-        ...state,
-        entities: {
-          ...state.entities,
-          [code]: newState
-        }
-      };
-    }
-
     case fromProduct.LOAD_PRODUCT_SUCCESS: {
       const detail = action.payload;
 
@@ -34,7 +17,7 @@ export function reducer(
         ...state,
         entities: {
           ...state.entities,
-          [detail.code]: { loading: false, value: detail }
+          [detail.code]: detail
         }
       };
     }
@@ -43,7 +26,7 @@ export function reducer(
         ...state,
         entities: {
           ...state.entities,
-          [action.payload.code]: { loading: false, value: null }
+          [action.payload.code]: {}
         }
       };
     }
